Added the meta.py file
